Backend/requirements.txt
fastapi>=0.115.6
uvicorn>=0.34.0
sqlalchemy>=2.0.36
alembic>=1.14.0
pydantic>=2.10.3
pydantic-settings>=2.7.0
passlib[bcrypt]
python-jose[cryptography]
python-multipart>=0.0.19
